# faster iandnot between bitmap containers and run containers