Rather than converting a run container to a bitmap container and then performing iandnot, call removeRange directly. 

```
name                                          old time/op  new time/op  delta
AndNot/inPlace=true/left=bitmap/right=run-10  6.71µs ±32%  3.08µs ±45%  -54.05%  (p=0.000 n=20+17)
```