@lemire 2024-04-02
This is good and I expect to merge. I will give it some time in case someone wants to review it further.