@neena 2024-04-04
@lemire I just copied this change over to a benchmarking tool I have that tests a prod workload that uses this function and it seems to break something -- could you hold on merging until I get to the bottom what's happening?